validator/query_node/requirements.txt
httpx~=0.27
python-dotenv~=1.0
redis~=5.0
asyncpg==0.29.0
colorama==0.4.6